moved unzip proof of concept to DataFiles class
